Sources moved inside cocoapods folder
