+++ lib/cyclonedx/cocoapods/bom_builder.rb
