+++ lib/cyclonedx/cocoapods/license.rb
