+++ lib/cyclonedx/cocoapods/pod.rb
