+++ lib/cyclonedx/cocoapods/search_engine.rb
