+++ lib/cyclonedx/cocoapods/spdx_licenses.rb
