+++ spec/spec_helper.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require "cyclonedx/cocoapods"
+require "cyclonedx/cocoapods/cli_runner"
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
